Core Data: Check for presence of entity config before testing plural form.
When auto-generating method names for various entities in the data system we
want to use manually-listed plural forms if they exist.

Previously, however, we've been assuming that when we search for an entity's
config that it exists. While this probably hasn't been a real source of bugs
it does present an opportunity for an invalid-type runtime exception.

In this patch we're verifying that the config exists before we access the
`plural` property, eliminating the opportunity for the rutnime crash.