Core Data: Import types before exporting them
Previously we've been exporting the `Context` and `Updatable` types
from a helper module and also using those types in the same file.
This represents a type violation since we're not importing those
types into the module's namespace.

In this patch we're also importing them before re-exporting them
so that the names are available as we expect.